Correctly calculate video size for portrait videos
When the webcam container is rectangular, the video size is sometimes not calculated correctly for portrait mode. I'm not sure if my implementation covers all situations correctly. For my application it was working as expected, even when the device is rotated. In my case, the webcam width and height attributes are always set to screen size.